Organizing and working through solving all situations that could occur
